feat: 4강 완료 및 InitialPoint와 맞추기
송지원 튜터님이 제공해주신 시작점과 비교하여 코드 수정

+++ Assets/ScriptableObject/Scripts/RangedAttackSO.cs
@@ -6,7 +6,7 @@ public class RangedAttackSO : AttackSO
 {
     [Header("Ranged Attack Info")]
     public string bulletNameTag;
-    public int duration; // 어느 시간동안 나가는지
+    public float duration; // 어느 시간동안 나가는지
     public float spread; // 랜덤으로 얼마나 퍼질지
     public int numberOfProjectilesPerShoot; // 한번 나갈 때 몇 개씩 나가는지
     public float multipleProjectilesAngle;
